refactor: created separate component for left block of car card
