+++ src/components/car-card-left-block/index.ts
@@ -1,0 +1,1 @@
+export { CarCardLeftBlock } from './car-card-left-block';
