+++ src/components/car-card-left-block/types.ts
@@ -1,0 +1,8 @@
+import { TCarsData } from '../../shared/types';
+
+export type TCarCardLeftBlock = {
+  data?: TCarsData;
+  onImageChange: (v: React.ChangeEvent<HTMLInputElement>) => void;
+  image: File | null;
+  imageUrl: string;
+};
